+++ Aplicación Final/Backend/demo/src/main/java/com/example/demo/model/Contenido.java
@@ -5,7 +5,7 @@ public class Contenido {
     private Long anho;
     private String descripcion;
     private String generosPertenece;
-    private Integer idContenido;
+    private Long idContenido;
     private String linkVideo;
     private String nombre;
     private String tipoContenido;
@@ -35,11 +35,11 @@ public class Contenido {
         this.generosPertenece= generosPertenece;
     }
 
-    public Integer getIdContenido(){
+    public Long getIdContenido(){
         return this.idContenido;
     }
 
-    public void setIdContenido(Integer idContenido){
+    public void setIdContenido(Long idContenido){
         this.idContenido= idContenido;
     }
